Made the code valid for a C++ compiler (void* stuff), plus some cleanup
git-svn-id: http://svn.xiph.org/trunk/speex@4060 0101bb08-14d6-0310-b084-bc0e0c8e3800

+++ libspeex/testenc_wb.c
@@ -90,7 +90,7 @@ int main(int argc, char **argv)
       }
       speex_bits_reset(&bits);
       for (i=0;i<FRAME_SIZE;i++)
-         in[i]=input[i];
+         in[i]=(short)input[i];
       for (i=0;i<FRAME_SIZE;i++)
          bak2[i]=bak[i];
       fwrite(in, sizeof(short), FRAME_SIZE, fout);
